pyproject.toml
[project]
name = "genvoy"
version = "0.1.0"
description = "Production-grade MCP server for fal.ai generative media workflows"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Genvoy Contributors" },
]
keywords = ["mcp", "model-context-protocol", "fal", "ai", "media-generation", "fastmcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "aiofiles>=25.1.0",
    "fastmcp>=3.0.2",
    "httpx>=0.27.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "respx>=0.22.0",
    "ruff>=0.15.2",
]

[project.scripts]
genvoy = "genvoy.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.11"
strict = false

[[tool.mypy.overrides]]
module = ["aiofiles"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
